fix(upval): thread the u64 thread-id domain end-to-end (codex HIGH)
The tagged-Cell shrink stored thread_id as u64 but the API boundary still
round-tripped through usize: UpVal::open / new_upval_open took usize,
try_open_payload returned usize, func.rs narrowed current_thread_id with
`as usize`, and every consumer re-widened with `as u64`. On wasm32
(usize = 32-bit, a supported embedding target) ids above u32::MAX truncated
at capture, so open-upvalue reads consulted the wrong thread/mirror and
unreachable-thread closing could reject on a truncated-id mismatch.

Use u64 throughout: open/new_upval_open take u64, try_open_payload returns
u64, and all consumers (vm.rs GET/SETUPVAL, state.rs upvalue_get/set +
unreachable-close, func.rs new_open_upval, coro_lib resume snapshots,
debug.rs upval-name) drop the round-trip casts. Size-neutral: the stored
field was already u64, so GcBox<UpVal> stays 48 B.

Tests: the u64 domain test now uses 0xFEDCBA9876543210 (> i64::MAX,
> u32::MAX) so it detects i64/u32 narrowing; adds the missing
closed_value()==Nil-on-open assertion (the byte-identical edge).

+++ crates/lua-vm/src/func.rs
@@ -58,7 +58,7 @@ fn new_open_upval(state: &mut LuaState, level: StackIdx, insert_pos: usize) -> G
     // executing `find_upval` — it captures one of that thread's stack slots.
     // `GlobalState::cross_thread_upvals` is what a coroutine actually reads
     // or writes through when accessing an upvalue belonging to its parent.
-    let owner_tid = state.global().current_thread_id as usize;
+    let owner_tid = state.global().current_thread_id;
     let uv: GcRef<UpVal> = state.new_upval_open(owner_tid, level);
     // Vec insert maintains descending StackIdx order (highest first),
     // mirroring the C intrusive list where the head is always the topmost slot.
